pyproject.toml
[tool.poetry]
name = "pubsub"
version = "0.1.0"
description = ""
authors = ["whiteking64"]

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-pubsub = "^2.26.0"
apache-beam = {extras = ["gcp", "test"], version = "^2.59.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
